refactor(QuickActionsModal): decompose 2,528-line monolith into directory module
Shell drops from 2,528 → 733 lines (−71%). Command logic extracted into 11
per-category builder files under commands/, two presentational components
promoted to QuickActionsList and QuickActionsSearchBar, and two pure utils
(activeTabInfo, quickActionSorting) isolated for independent testing.

Barrel at QuickActionsModal/index.ts preserves the existing named-export
import path used by AppUtilityModals.tsx. Behaviour is byte-equivalent —
same keyboard nav, mode transitions (main/move-to-group/agents), LIVE/IDLE
bucket headers, rename-in-place flow, and revealJumpTarget logic.

Existing 147-test integration suite passes unchanged via the barrel.
17 new unit tests added across 7 files covering command builders, components,
and utils. Full suite: 29,229 passing / 0 failing.
