+++ src/renderer/components/QuickActionsModal/index.ts
@@ -1,0 +1,1 @@
+export { QuickActionsModal } from './QuickActionsModal';
